+++ Algorithm.Framework/Portfolio/SimplePortfolioConstructionModel.cs
@@ -58,7 +58,7 @@ namespace QuantConnect.Algorithm.Framework.Portfolio
         /// <param name="changes">The security additions and removals from the algorithm</param>
         public void OnSecuritiesChanged(QCAlgorithmFramework algorithm, SecurityChanges changes)
         {
-            NotifiedSecurityChanged.UpdateCollection(_securities, changes);
+            NotifiedSecurityChanges.UpdateCollection(_securities, changes);
         }
     }
 }
